+++ CHANGELOG.md
@@ -5,7 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## 1.2.0
+## 1.2.0-9-3
+
+### Added
+
+- Added date_created to check in data log
+
+### Removed
+
+- Removed Slack post from logError saga
+
+## 1.2.0-9-2
 
 ### Added